requirement.txt
absl-py==1.3.0
addict==2.4.0
appdirs==1.4.4
asttokens==2.0.8
attrs==22.1.0
audioread==3.0.0
av==9.2.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@3702849800aa56e2223035bccd1c6ef91c704ca8
colorama==0.4.5
coloredlogs==15.0.1
comm==0.1.2
contourpy==1.0.5
coverage==6.5.0
cycler==0.11.0
debugpy==1.6.3
decorator==4.4.2
decord==0.6.0
dill==0.3.5.1
einops==0.6.0
etils==0.8.0
executing==1.1.0
filelock==3.8.0
flake8==5.0.4
flatbuffers==22.9.24
fonttools==4.37.4
ftfy==6.1.1
future==0.18.2
google-auth==2.12.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.56.4
grpcio==1.49.1
huggingface-hub==0.10.0
humanfriendly==10.0
idna==3.4
imageio==2.22.1
imageio-ffmpeg==0.4.7
imgaug==0.4.0
importlib-metadata==5.0.0
importlib-resources==5.9.0
iniconfig==1.1.1
interrogate==1.5.0
ipdb==0.13.9
ipykernel==6.16.0
ipython==8.5.0
isort==4.3.21
jedi==0.18.1
joblib==1.2.0
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.4
librosa==0.9.2
llvmlite==0.39.1
lmdb==1.3.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mmaction2==0.18.0
mmcv-full==1.3.12
moviepy==1.0.3
mpmath==1.2.1
nest-asyncio==1.5.6
networkx==2.8.7
numba==0.56.2
numpy==1.23.3
oauthlib==3.2.1
onnx==1.12.0
onnxruntime==1.12.1
opencv-contrib-python==4.6.0.66
opencv-python==4.5.3.56
opencv-python-headless==4.2.0.34
packaging==21.3
pandas==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
PIMS==0.6.1
platformdirs==3.0.0
pluggy==1.0.0
pooch==1.6.0
proglog==0.1.10
promise==2.3
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.1.3
pytest-runner==6.0.0
python-dateutil==2.8.2
PyTurboJPEG==1.6.7
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
Shapely==1.8.4
six==1.16.0
slicerator==1.1.0
soundfile==0.11.0
stack-data==0.5.1
sympy==1.11.1
tabulate==0.8.10
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-datasets==4.6.0
tensorflow-metadata==1.10.0
termcolor==2.0.1
threadpoolctl==3.1.0
tifffile==2022.8.12
timm==0.6.11
toml==0.10.2
tomli==2.0.1
torch==1.10.0+cu113
torch-tb-profiler==0.4.1
torchaudio==0.10.0+cu113
torchvision==0.11.1+cu113
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
xdoctest==1.1.0
yacs==0.1.8
yapf==0.32.0
zipp==3.8.1
